samples/langgraph-sql-agent/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "langgraph-sql-agent"
version = "0.1.0"
description = "A LangGraph ReAct agent that can run queries on an ephemeral SQLite database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.16",
    "langchain-google-vertexai>=2.0.7",
    "langgraph>=0.4.3",
    "opentelemetry-exporter-gcp-logging>=1.9.0a0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.0",
    "opentelemetry-instrumentation-httpx>=0.54b0",
    "opentelemetry-instrumentation-requests>=0.54b0",
    "opentelemetry-instrumentation-sqlite3>=0.54b0",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "rich>=14.0.0",
]
